Add advertisement management schema and entities
This commit introduces a new database schema for managing advertisements in the AqarakDB application. It includes the creation of tables such as Ads, FavoriteAds, ImageAds, Offers, PropertyContractPhotos, SaveAds, and VideoAds under the ads and offers schemas. Migration files define the structure and relationships of these tables with existing entities like Users, Subscriptions, and Payments. The AppDbContext class is updated to include DbSet properties for the new entities, ensuring seamless data access and manipulation. Model snapshots and migration files maintain synchronization between the database schema and the application's data model.